Refactor mathematical functions into seperate package
This will serve as a header only templated package to perform several commonly used mathematical operations. Physics may be refactored into its own package should the need to further expand it arise.
