--- src/coordinates.cpp
@@ -1,1 +1,0 @@
-#include <coordinates.h>
