+++ src/math/physics.h
@@ -1,0 +1,13 @@
+#pragma once
+#include <math/vector3d.h>
+
+namespace rpg::math {
+template <typename T>
+concept Kinematic = requires {
+  { T::pos }
+  ->Vector3D;
+  { T::vel }
+  ->Vector3D;
+};
+
+}  // namespace rpg::math
